Dedup wordpressAdapter
